refactor(api): move archivecampaigns route to BC
