+++ api/src/prescription/campaign/application/campaign-adminstration-controller.js
@@ -4,6 +4,7 @@ import * as checkAdminMemberHasRoleSuperAdminUseCase from '../../../../lib/appli
 import * as csvSerializer from '../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
 import * as requestResponseUtils from '../../../../lib/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../../../../src/prescription/campaign/domain/usecases/index.js';
+import * as csvCampaignsIdsParser from '../infrastructure/serializers/csv/csv-campaigns-ids-parser.js';
 import * as campaignReportSerializer from '../infrastructure/serializers/jsonapi/campaign-report-serializer.js';
 
 const createCampaigns = async function (request, h, dependencies = { csvSerializer }) {
@@ -99,6 +100,17 @@ const unarchiveCampaign = function (request, h, dependencies = { campaignReportS
   return usecases.unarchiveCampaign({ userId, campaignId }).then(dependencies.campaignReportSerializer.serialize);
 };
 
+const archiveCampaigns = async function (request, h, dependencies = { csvCampaignsIdsParser }) {
+  const { userId } = request.auth.credentials;
+  const campaignIds = await dependencies.csvCampaignsIdsParser.extractCampaignsIds(request.payload.path);
+
+  await usecases.archiveCampaigns({
+    userId,
+    campaignIds,
+  });
+
+  return h.response(null).code(204);
+};
 const campaignAdministrationController = {
   save,
   update,
@@ -107,6 +119,7 @@ const campaignAdministrationController = {
   updateCampaignDetails,
   updateCampaignCode,
   archiveCampaign,
+  archiveCampaigns,
   unarchiveCampaign,
 };
