+++ api/src/prescription/campaign/infrastructure/serializers/csv/campaign-ids-csv-header.js
@@ -1,4 +1,4 @@
-import { CsvColumn } from '../csv-column.js';
+import { CsvColumn } from '../../../../../../lib/infrastructure/serializers/csv/csv-column.js';
 
 class CampaignIdsCsvHeader {
   constructor() {
