--- api/tests/unit/application/campaigns-administration/index_test.js
@@ -1,25 +1,0 @@
-import { campaignController } from '../../../../lib/application/campaigns-administration/campaign-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/campaigns-administration/index.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Unit | Application | Router | campaign-participation-router ', function () {
-  describe('PATCH /api/admin/campaigns/archive-campaigns', function () {
-    it('returns 200 when admin member has rights', async function () {
-      // given
-      sinon.stub(campaignController, 'archiveCampaigns').returns(null);
-      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      await httpTestServer.request('POST', '/api/admin/campaigns/archive-campaigns', {});
-
-      // then
-      expect(securityPreHandlers.hasAtLeastOneAccessOf).to.have.been.calledWithExactly([
-        securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-        securityPreHandlers.checkAdminMemberHasRoleMetier,
-      ]);
-    });
-  });
-});
